correcao_atividades-assignments/01-programming/02-control-flow

+++ assignments/01-programming/02-control-flow/06-heart/submissions/luiz-eduardo-s-carneiro/src/HeartRates.java
@@ -70,8 +70,8 @@ public class HeartRates{
 
     public String calculateTargetHeartRate(){
         int max = calculateMaxHeartRate();
-        int minTarget = max * 0.50;
-        int maxTarget = max * 0.85;
+        double minTarget = max * 0.50;
+        double maxTarget = max * 0.85;
 
         return String.format("%d bpm - %d bpm", minTarget, maxTarget);
     }
